feat: add lorem text and headers
