+++ src/index.html
@@ -15,6 +15,17 @@
   <div id="app"></div>
   <div id="chart"></div>
   <div id="map">
+    <div id="title">
+      <h1>Washington D.C. At Risk</h1>
+      <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Quisquam voluptatum qui aliquid ratione facere
+        necessitatibus vero, maxime nam reprehenderit vitae et repellendus aut ab praesentium nulla, fugit consectetur
+        eveniet illum.</p>
+      <p>Eum excepturi odit laborum, quis harum animi hic explicabo incidunt non similique sint aut est iure sunt
+        tenetur totam provident quae sequi maxime. Magnam veritatis, dolorum reiciendis nesciunt alias temporibus.</p>
+      <p>Itaque dolorum optio quaerat eaque, hic laudantium, quo, tempora sint praesentium aliquam animi assumenda neque
+        ipsum. Quod autem ipsum sunt! Sed dignissimos ipsa facilis, pariatur asperiores labore cupiditate quae
+        temporibus!</p>
+    </div>
     <div id="legend">
       <!-- <button id="changeMap">Change Map</button> -->
       <h4>Map Type</h4>
