+++ src/styles.css
@@ -29,7 +29,8 @@ body {
 
 #map {
   height: 680px;
-  border: 1px solid #999;
+  /* border: 1px solid #999; */
+  max-width: 1000px;
 }
 
 /* tooltip styling */
@@ -83,6 +84,19 @@ body {
   border-bottom: 1px solid #aaa;
 }
 
+#title {
+  position: absolute;
+  border: 2px solid #ccc;
+  top: 60px;
+  left: 20px;
+  width: 300px;
+  height: 350px;
+  background-color: #fff;
+  z-index: 1000;
+  padding: 4px 16px;
+  opacity: 0.8;
+}
+
 #legend {
   /* font-size: .875em; */
   position: absolute;
@@ -94,6 +108,7 @@ body {
   z-index: 1000;
   background-color: #eee;
   padding: 4px 16px;
+  opacity: 1;
 }
 
 #legend h4 {
