+++ src/utils.js
@@ -176,10 +176,83 @@ export const buildLeafletMap = (regions, width, height) => {
     .addTo(map)
 
   // fit bounds of polygon and get map view to fit (thank God for this method!)
-  map.fitBounds(WashingtonDCLayer.getBounds())
+  map.fitBounds(WashingtonDCLayer.getBounds(), { paddingTopLeft: [400, 0] })
   console.log('regions.getBounds()', WashingtonDCLayer.getBounds())
   return {
     map,
     mapLayer,
   }
+}
+
+export const addLegendEventListeners = (map, mapLayer) => {
+  let tempMapLayer = mapLayer
+
+  const radioHTMLCol = document.getElementsByName('mapType')
+  const radioArr = [...radioHTMLCol]
+  radioArr.forEach((item) => {
+    item.addEventListener('change', (e) => {
+      switch (parseInt(+e.target.value, 10)) {
+        case 1:
+          map.removeLayer(tempMapLayer)
+          tempMapLayer = L.tileLayer(
+            'https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png',
+            {
+              maxZoom: 15,
+              attribution:
+                '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors',
+            }
+          ).addTo(map)
+          break
+        case 2:
+          map.removeLayer(tempMapLayer)
+          tempMapLayer = L.tileLayer(
+            'https://tiles.stadiamaps.com/tiles/alidade_smooth/{z}/{x}/{y}{r}.png',
+            {
+              maxZoom: 15,
+              attribution:
+                '&copy; <a href="https://stadiamaps.com/">Stadia Maps</a>, &copy; <a href="https://openmaptiles.org/">OpenMapTiles</a> &copy; <a href="http://openstreetmap.org">OpenStreetMap</a> contributors',
+            }
+          ).addTo(map)
+          break
+        case 3:
+          map.removeLayer(tempMapLayer)
+          tempMapLayer = L.tileLayer(
+            'https://tiles.stadiamaps.com/tiles/outdoors/{z}/{x}/{y}{r}.png',
+            {
+              maxZoom: 20,
+              attribution:
+                '&copy; <a href="https://stadiamaps.com/">Stadia Maps</a>, &copy; <a href="https://openmaptiles.org/">OpenMapTiles</a> &copy; <a href="http://openstreetmap.org">OpenStreetMap</a> contributors',
+            }
+          ).addTo(map)
+          break
+        case 4:
+          map.removeLayer(tempMapLayer)
+          tempMapLayer = L.tileLayer(
+            'https://server.arcgisonline.com/ArcGIS/rest/services/World_Imagery/MapServer/tile/{z}/{y}/{x}',
+            {
+              attribution:
+                'Tiles &copy; Esri &mdash; Source: Esri, i-cubed, USDA, USGS, AEX, GeoEye, Getmapping, Aerogrid, IGN, IGP, UPR-EGP, and the GIS User Community',
+            }
+          ).addTo(map)
+          break
+        case 5:
+          map.removeLayer(tempMapLayer)
+          tempMapLayer = L.tileLayer(
+            'https://stamen-tiles-{s}.a.ssl.fastly.net/terrain/{z}/{x}/{y}{r}.{ext}',
+            {
+              attribution:
+                'Map tiles by <a href="http://stamen.com">Stamen Design</a>, <a href="http://creativecommons.org/licenses/by/3.0">CC BY 3.0</a> &mdash; Map data &copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors',
+              subdomains: 'abcd',
+              minZoom: 0,
+              maxZoom: 18,
+              ext: 'png',
+            }
+          ).addTo(map)
+          break
+        default:
+      }
+    })
+  })
+  console.log(`radioHTMLCol`, radioHTMLCol)
+  console.log(`radioArr`, radioArr)
 }
